feat(NodeFeature): split components
