--- components/Home/NodeFeature/index.module.scss
@@ -1,31 +1,0 @@
-.nodeFeatures {
-  display: grid;
-  grid-gap: 1.5rem;
-  grid-template-columns: repeat(3, 1fr);
-  justify-content: center;
-  margin: 0 auto 5em auto;
-
-  @media (max-width: 900px) {
-    grid-template-columns: 1fr;
-  }
-
-  .text {
-    margin: 0 2em;
-    max-width: 248px;
-    text-align: center;
-
-    img {
-      width: 65%;
-    }
-
-    .featureIcon {
-      height: auto;
-      min-width: 120px;
-      opacity: 0.75;
-
-      @media (max-width: 900px) {
-        min-width: 60px;
-      }
-    }
-  }
-}
