+++ components/Home/NodeFeatures/NodeFeature/index.module.scss
@@ -1,0 +1,5 @@
+.container {
+  margin: 0 2em;
+  max-width: 248px;
+  text-align: center;
+}
